Rename directory 3dModel -> 3dModelViewer
